ml/requirements.txt
pandas==2.2.3
numpy==2.0.2
pyarrow==18.1.0
scikit-learn==1.5.2
xgboost==2.1.3
requests==2.32.3
tqdm==4.67.1
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
python-dotenv==1.0.1
joblib==1.4.2
